Move ContinuableScope and Continuation inside ContinuableScopeManager
And make private.
